app/pyproject.toml
[project]
name = "forensiq-backend"
version = "0.1.0"
description = "Backend service for the ForensIQ Application Stack"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.13.1",
    "boto3>=1.34.0",
    "bcrypt==4.0.1",
    "cryptography>=42.0.0",
    "elasticsearch[async]>=8.13.0",
    "fastapi==0.110.0",
    "passlib[bcrypt]==1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings==2.2.1",
    "pydantic[email]>=2.3.0",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.22",
    "sqlalchemy>=2.0.30",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "uvicorn[standard]==0.27.1",
    "river>=0.21.0",
    "networkx>=3.0",
    "pyyaml>=6.0",
    "numpy>=1.26.0",
]
